requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
pydantic
cachetools
slowapi
gunicorn
